North direction added; Maps and Buildings moved to separate module SceneSurroundings
