+++ packages/viewer/src/assets/example.html
@@ -47,9 +47,10 @@
           value="51.506810732490656" />
         <input id="zeroCoordInputY" type="text" name="zeroCoordY" placeholder="Yval" style="width:24.25%"
           value="-0.0892642750895124" />
-        <select id="mapUnits" style="width: 12%; height: 40px;" onchange="v.addMap()"></select>
-        <select id="providerColor" style="width: 24.25%; height: 40px;" onchange="v.addMap()"></select>
-        <button class="" onclick="v.removeMap(),v.hideBuild()" style="width:12%">Remove maps</button>
+        <input id="North angle" type="text" name="NorthAngle" placeholder="angle" style="width:12%" value="0" />
+        <select id="mapUnits" style="width: 12%; height: 40px;" onchange="v.addMapAndBuild()"></select>
+        <select id="providerColor" style="width: 12%; height: 40px;" onchange="v.addMapAndBuild()"></select>
+        <button class="" onclick="v.removeMapAndBuild()" style="width:12%">Remove maps</button>
       </div>
       <div class="twelve columns">
         <button onclick="v.unloadAll()">Dispose Everything</button>
